DRAFT: CLE specification (#3)
Continuing on from our PR
[here](https://github.com/Ecma-TC54/tg3/pull/2) which added CLE
constraints, we're now adding the specification.

> The Rationale (this PR). The rationale document describes the goals
and constraints of the specification.
> The Technical Choices. The technical choices should be very easy to
understand given the constraints and goals defined in the Rationale
document.
> An Example Spec. Once we have the constraints and the technical
choices, we can make an example specification.

We were supposed to do a Technical Choices document, but I found that it
was somewhat unnecessary. Happy to add it though if we feel like its
needed.

There are some places in this specification where I have a `TODO` which
I would like to discuss in the tg3 wg

Original CLE google doc is
[here](https://docs.google.com/document/d/1sRMS1IX0r7ZkYthDR0VY1bYyvp_6K_xw4sR1vZwla8E/edit).

---------

Signed-off-by: Benji Visser <benji@093b.org>
Co-authored-by: Jordan Harband <ljharb@gmail.com>